Revert filesystem client tests to devel (#814)
